STLport-ms-compat

+++ test/eh/random_number.cpp
@@ -1,6 +1,6 @@
 /***********************************************************************************
-	random_number.cpp
-	
+  random_number.cpp
+
  * Copyright (c) 1997
  * Mark of the Unicorn, Inc.
  *
@@ -23,13 +23,13 @@
 # include <stdlib.h>
 #endif
 
-unsigned random_number( unsigned range )
+unsigned random_number( size_t range )
 {
 #if !defined( __SGI_STL )
-	if (range == 0) return 0;
-	return (unsigned)(EH_STD::rand() + EH_STD::rand()) % range;
+  if (range == 0) return 0;
+  return (unsigned)(EH_STD::rand() + EH_STD::rand()) % range;
 #else
-	static EH_STD::subtractive_rng rnd;
+  static EH_STD::subtractive_rng rnd;
         if (range==0) return 0;
         return rnd(range);
 #endif
